backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pydantic==2.5.1
playwright==1.41.2
aiohttp==3.9.3
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.6
redis==5.0.1
lxml==5.1.0
asyncpg==0.29.0 